pyproject.toml
[tool.poetry]
name = "ask-cli"
version = "0.1.0"
description = ""
authors = ["Raphael Obadia <raphael.obadia@loreal.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ruff = "^0.3.5"
black = "^24.3.0"
typer = "^0.12.1"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
google-auth = "^2.29.0"
requests = "^2.31.0"
openai = "^1.16.2"
halo = "^0.0.31"
google-generativeai = "^0.5.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.5"
black = "^24.3.0"
pytest = "^8.1.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191", # indentation contains tabs
]

